backend/requirements.txt

setuptools
undetected-chromedriver>=3.5.0
selenium>=4.15.0
requests>=2.31.0

# FastAPI and web service
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.1.0
soundcard>=0.4.2,<0.5.0
soundfile>=0.12.1,<0.13.0
numpy>=1.21.0,<2.0.0

# AI and messaging
google-generativeai>=0.3.0
httpx>=0.24.0

# Telegram bot dependencies
aiogram>=3.0.0
aiohttp>=3.8.0
python-dotenv>=1.0.0
python-dotenv>=1.0.0


python-multipart  # For file uploads if needed
